+++ CHANGES.md
@@ -81,6 +81,9 @@ next
 - Rename `files_recursively_in` to `source_tree` to make it clearer it
   doesn't include generated files (#899, fix #843, @diml)
 
+- Present the `menhir` stanza as an extension with its own version
+  (#..., @diml)
+
 1.0+beta20 (10/04/2018)
 -----------------------